# request-service/internal/docs/swagger.json
{
    "components": {
        "schemas": {
            "httpx.ErrorBody": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "requests.createRequestBody": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "requests.quantityBody": {
                "properties": {
                    "quantity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "requests.requestExistsBody": {
                "properties": {
                    "existing": {
                        "$ref": "#/components/schemas/requests.requestResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "requests.requestResponse": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "createdBy": {
                        "description": "The customer who opened the request. Absent when nobody did - a request opened for\na seller offering against an item has no buyer behind it.",
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "exact": {
                        "description": "Set when this request carries the searched-for name outright rather than merely\nsomething like it. It is what tells a client that creating is not on offer here -\nonly joining is - so the form can say so before the customer finds out from a 409.",
                        "type": "boolean"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "requestId": {
                        "type": "string"
                    },
                    "similarity": {
                        "description": "How close this request's item name is to the one that was searched for, 0 to 1.\nCarried only by the near-match endpoints, so the caller can order and explain the\nsuggestions rather than being handed an unranked list.",
                        "type": "number"
                    },
                    "status": {
                        "type": "string"
                    },
                    "totalCustomers": {
                        "type": "integer"
                    },
                    "totalOffers": {
                        "description": "How many live offers stand on this request. Counted by offer-service - PENDING or\nAPPROVED, since a cancelled or rejected one is a record rather than an offer - and\nreported here because it is half of what the status means.",
                        "type": "integer"
                    },
                    "totalQuantity": {
                        "type": "integer"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "description": "Paste the accessToken from POST /api/auth/login.",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Purchase requests and the aggregated demand behind them. Public API only: the /internal endpoints are service-to-service and have no route through the gateway.",
        "title": "request-service",
        "version": "1.0.0"
    },
    "externalDocs": {
        "description": "",
        "url": ""
    },
    "paths": {
        "/api/requests": {
            "get": {
                "description": "Open to everyone, signed in or not. It carries aggregate totals and no\nparticipant identity, so there is nothing here a token would protect.",
                "parameters": [
                    {
                        "description": "Match on item name",
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by category",
                        "in": "query",
                        "name": "category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "OPEN or INACTIVE",
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Page size, 1-100",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Rows to skip",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/requests.requestResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad paging parameters"
                    }
                },
                "summary": "Browse open demand",
                "tags": [
                    "requests"
                ]
            },
            "post": {
                "description": "Opens a request other customers can join. The caller becomes its first\nparticipant; the customerId is resolved from the token, never sent.\nIf a request already carries this item name, nothing is created: the 409\ncomes back with that request attached, and joining it is the\ncustomer's own call through the participants endpoint. That holds for a\nrequest with no buyers on it too - one a seller opened by offering\nagainst the item, or one everybody has left - because joining is what\nmakes it demand again. A merely similar\nname is not refused - see /api/requests/similar, which is what the\nnew-request form shows while the name is being typed.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/requests.createRequestBody",
                                        "summary": "body",
                                        "description": "Item, category and the caller's own quantity"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Item, category and the caller's own quantity",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/requests.requestResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Validation failed"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a CUSTOMER, or no customer profile"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/requests.requestExistsBody"
                                }
                            }
                        },
                        "description": "Already a participant, or this item already has a request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create a purchase request",
                "tags": [
                    "requests"
                ]
            }
        },
        "/api/requests/me": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/requests.requestResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a CUSTOMER"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List the caller's own requests",
                "tags": [
                    "requests"
                ]
            }
        },
        "/api/requests/similar": {
            "get": {
                "description": "Powers the suggestions a customer sees while naming an item, and is the\nsame match a create is refused on. Open to everyone, signed in or not:\nit returns the browse projection, ranked by similarity. Requests with no\nbuyers on them are included and carry their status, because those are\noften the ones most worth joining - a seller may already have offered\nagainst the item.",
                "parameters": [
                    {
                        "description": "The item name being typed",
                        "in": "query",
                        "name": "itemName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/requests.requestResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "itemName is required"
                    }
                },
                "summary": "Find requests whose item name is close to this one",
                "tags": [
                    "requests"
                ]
            }
        },
        "/api/requests/{requestId}": {
            "get": {
                "description": "Open to everyone, signed in or not. Returns totals, not the participant\nlist: who joined is withheld from the public projection and served only\non the internal endpoint.",
                "parameters": [
                    {
                        "description": "Request id",
                        "in": "path",
                        "name": "requestId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/requests.requestResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Malformed id"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Read one request with its aggregated demand",
                "tags": [
                    "requests"
                ]
            }
        },
        "/api/requests/{requestId}/participants": {
            "post": {
                "description": "Adds the caller's demand to the aggregate. Joining twice is refused;\nuse the participants/me endpoint to change a quantity.",
                "parameters": [
                    {
                        "description": "Request id",
                        "in": "path",
                        "name": "requestId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/requests.quantityBody",
                                        "summary": "body",
                                        "description": "Quantity wanted"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Quantity wanted",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/requests.requestResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Quantity must be positive"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a CUSTOMER"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Already joined, or the request is not OPEN"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Join a request with a quantity",
                "tags": [
                    "participation"
                ]
            }
        },
        "/api/requests/{requestId}/participants/me": {
            "delete": {
                "parameters": [
                    {
                        "description": "Request id",
                        "in": "path",
                        "name": "requestId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Withdrawn"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a CUSTOMER"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a participant"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "The request is no longer OPEN"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Withdraw from a request",
                "tags": [
                    "participation"
                ]
            },
            "put": {
                "description": "Addressed as participants/me rather than by participant id: a customer\ncan only ever change their own row, so there is nothing else to name.",
                "parameters": [
                    {
                        "description": "Request id",
                        "in": "path",
                        "name": "requestId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/requests.quantityBody",
                                        "summary": "body",
                                        "description": "New quantity"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "New quantity",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/requests.requestResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a CUSTOMER"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "Not a participant"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorBody"
                                }
                            }
                        },
                        "description": "The request is no longer OPEN"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Change how much the caller wants",
                "tags": [
                    "participation"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}